+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.22.1](https://github.com/kinguru-io/kinguru-next/compare/v0.22.0...v0.22.1) (2024-02-13)
+
+
+### Bug Fixes
+
+* pin react query library version ([b2f9d43](https://github.com/kinguru-io/kinguru-next/commit/b2f9d4327f7cb8c5801da1fcff3027817af58297))
+
+# [0.22.0](https://github.com/kinguru-io/kinguru-next/compare/v0.21.0...v0.22.0) (2024-02-11)
+
+
+### Features
+
+* **KINGU-77:** auth page layout ([#418](https://github.com/kinguru-io/kinguru-next/issues/418)) ([03191a1](https://github.com/kinguru-io/kinguru-next/commit/03191a13fd99d2f24fecbb661a2ebaa982d50379))
+
 # [0.21.0](https://github.com/kinguru-io/kinguru-next/compare/v0.20.1...v0.21.0) (2024-02-06)